[web] Move products loading to SoftwarerProvider
